modify test_conditional
